Ramya-Nachiket/requirements.txt
beautifulsoup4==4.10.0
certifi==2021.10.8
charset-normalizer==2.0.8
click==8.0.3
cloudpickle==2.0.0
colorama==0.4.4
cycler==0.11.0
dask==2021.11.2
dask-jobqueue==0.7.3
distributed==2021.11.2
docopt==0.6.2
fonttools==4.28.2
fsspec==2021.11.0
HeapDict==1.0.1
idna==3.3
Jinja2==3.0.3
joblib==1.1.0
Js2Py==0.71
kiwisolver==1.3.2
locket==0.2.1
MarkupSafe==2.0.1
matplotlib==3.5.0
msgpack==1.0.3
multipledispatch==0.6.0
numpy==1.21.4
packaging==21.3
pandas==1.3.4
partd==1.2.0
Pillow==8.4.0
pipwin==0.5.1
psutil==5.8.0
PyAudio @ file:///C:/Users/ADMIN/pipwin/PyAudio-0.2.11-cp310-cp310-win_amd64.whl
pyjsparser==2.7.1
pyparsing==3.0.6
PyPrind==2.11.3
pySmartDL==1.3.4
python-dateutil==2.8.2
pytz==2021.3
pytz-deprecation-shim==0.1.0.post0
PyYAML==6.0
requests==2.26.0
scikit-learn==1.0.1
scipy==1.7.3
setuptools-scm==6.3.2
six==1.16.0
sklearn==0.0
sortedcontainers==2.4.0
soupsieve==2.3.1
tblib==1.7.0
threadpoolctl==3.0.0
tomli==1.2.2
toolz==0.11.2
tornado==6.1
typing_extensions==4.0.0
tzdata==2021.5
tzlocal==4.1
urllib3==1.26.7
zict==2.0.0
